+++ crates/libspot/src/peaks.rs
@@ -4,7 +4,7 @@
 //! about peaks data using an underlying Ubend circular buffer.
 
 use crate::error::SpotResult;
-use crate::Float;
+
 use crate::math::is_nan;
 use crate::ubend::Ubend;
 
@@ -12,13 +12,13 @@ use crate::ubend::Ubend;
 #[derive(Debug, Clone)]
 pub struct Peaks {
     /// Sum of the elements
-    e: Float,
+    e: f64,
     /// Sum of the square of the elements
-    e2: Float,
+    e2: f64,
     /// Minimum of the elements
-    min: Float,
+    min: f64,
     /// Maximum of the elements
-    max: Float,
+    max: f64,
     /// Underlying data container
     container: Ubend,
 }
@@ -29,8 +29,8 @@ impl Peaks {
         Ok(Self {
             e: 0.0,
             e2: 0.0,
-            min: f64::NAN as Float,
-            max: f64::NAN as Float,
+            min: f64::NAN,
+            max: f64::NAN,
             container: Ubend::new(size)?,
         })
     }
@@ -41,7 +41,7 @@ impl Peaks {
     }
 
     /// Add a new data point into the peaks
-    pub fn push(&mut self, x: Float) {
+    pub fn push(&mut self, x: f64) {
         let erased = self.container.push(x);
         let size = self.size();
 
@@ -72,20 +72,20 @@ impl Peaks {
     }
 
     /// Compute the mean of the elements
-    pub fn mean(&self) -> Float {
+    pub fn mean(&self) -> f64 {
         let size = self.size();
         if size == 0 {
-            f64::NAN as Float
+            f64::NAN
         } else {
             self.e / (size as f64)
         }
     }
 
     /// Compute the variance of the elements
-    pub fn variance(&self) -> Float {
+    pub fn variance(&self) -> f64 {
         let size = self.size();
         if size == 0 {
-            f64::NAN as Float
+            f64::NAN
         } else {
             let size_f = size as f64;
             let mean = self.e / size_f;
@@ -94,22 +94,22 @@ impl Peaks {
     }
 
     /// Get the minimum value
-    pub fn min(&self) -> Float {
+    pub fn min(&self) -> f64 {
         self.min
     }
 
     /// Get the maximum value
-    pub fn max(&self) -> Float {
+    pub fn max(&self) -> f64 {
         self.max
     }
 
     /// Get the sum of elements
-    pub fn sum(&self) -> Float {
+    pub fn sum(&self) -> f64 {
         self.e
     }
 
     /// Get the sum of squares
-    pub fn sum_squares(&self) -> Float {
+    pub fn sum_squares(&self) -> f64 {
         self.e2
     }
 
@@ -120,14 +120,14 @@ impl Peaks {
 
     /// Update all statistics by iterating through the container
     /// This is called when we need to recompute min/max after an erasure
-    /// 
+    ///
     /// NOTE: This implementation matches a bug in the C libspot implementation
     /// where e2 is incorrectly assigned instead of accumulated during stats update.
     /// This maintains exact compatibility with the C library results.
     fn update_stats(&mut self) {
         // Reset min and max
-        self.min = f64::NAN as Float;
-        self.max = f64::NAN as Float;
+        self.min = f64::NAN;
+        self.max = f64::NAN;
         // Reset accumulators
         self.e = 0.0;
         self.e2 = 0.0;
@@ -138,14 +138,14 @@ impl Peaks {
             // Direct access to container data (matches C implementation)
             let value = self.container.raw_data()[i];
             self.e += value;
-            
+
             // COMPATIBILITY NOTE: The C implementation has a bug here:
             // In peaks.c line 45: `peaks->e2 = value * value;` should be `peaks->e2 += value * value;`
             // We match this bug to ensure exact C compatibility as requested by the user.
             // This affects variance calculation which impacts MoM estimator and GPD parameter selection.
             // The C code overwrites e2 on every iteration, so only the last value's square remains.
             self.e2 = value * value;
-            
+
             if is_nan(self.min) || (value < self.min) {
                 self.min = value;
             }
@@ -184,7 +184,7 @@ mod tests {
     #[test]
     fn test_peaks_single_element() {
         let mut peaks = Peaks::new(3).unwrap();
-        
+
         peaks.push(5.0);
         assert_eq!(peaks.size(), 1);
         assert_relative_eq!(peaks.sum(), 5.0);
@@ -198,45 +198,45 @@ mod tests {
     #[test]
     fn test_peaks_multiple_elements() {
         let mut peaks = Peaks::new(5).unwrap();
-        
+
         peaks.push(1.0);
         peaks.push(2.0);
         peaks.push(3.0);
-        
+
         assert_eq!(peaks.size(), 3);
         assert_relative_eq!(peaks.sum(), 6.0);
         assert_relative_eq!(peaks.sum_squares(), 14.0);
         assert_relative_eq!(peaks.min(), 1.0);
         assert_relative_eq!(peaks.max(), 3.0);
         assert_relative_eq!(peaks.mean(), 2.0);
-        
+
         // Variance = E[X²] - (E[X])² = 14/3 - 4 = 14/3 - 12/3 = 2/3
-        assert_relative_eq!(peaks.variance(), 2.0/3.0, epsilon = 1e-14);
+        assert_relative_eq!(peaks.variance(), 2.0 / 3.0, epsilon = 1e-14);
     }
 
     #[test]
     fn test_peaks_overflow_and_min_max_update() {
         let mut peaks = Peaks::new(3).unwrap();
-        
+
         // Fill with 1, 2, 3
-        peaks.push(1.0);  // min=1, max=1
-        peaks.push(2.0);  // min=1, max=2
-        peaks.push(3.0);  // min=1, max=3
-        
+        peaks.push(1.0); // min=1, max=1
+        peaks.push(2.0); // min=1, max=2
+        peaks.push(3.0); // min=1, max=3
+
         assert_relative_eq!(peaks.min(), 1.0);
         assert_relative_eq!(peaks.max(), 3.0);
-        
+
         // Add 0.5, which should erase 1.0 and become new minimum
-        peaks.push(0.5);  // should erase 1.0, so we have [2, 3, 0.5]
-        
+        peaks.push(0.5); // should erase 1.0, so we have [2, 3, 0.5]
+
         assert_eq!(peaks.size(), 3);
         assert_relative_eq!(peaks.min(), 0.5);
         assert_relative_eq!(peaks.max(), 3.0);
         assert_relative_eq!(peaks.sum(), 5.5);
-        
+
         // Add 4.0, which should erase 2.0 and become new maximum
-        peaks.push(4.0);  // should erase 2.0, so we have [3, 0.5, 4.0]
-        
+        peaks.push(4.0); // should erase 2.0, so we have [3, 0.5, 4.0]
+
         assert_relative_eq!(peaks.min(), 0.5);
         assert_relative_eq!(peaks.max(), 4.0);
         assert_relative_eq!(peaks.sum(), 7.5);
@@ -245,24 +245,24 @@ mod tests {
     #[test]
     fn test_peaks_stats_after_min_erasure() {
         let mut peaks = Peaks::new(3).unwrap();
-        
+
         // Add values where the minimum will be erased
         peaks.push(2.0);
-        peaks.push(1.0);  // This is the minimum
+        peaks.push(1.0); // This is the minimum
         peaks.push(3.0);
-        
+
         assert_relative_eq!(peaks.min(), 1.0);
         assert_relative_eq!(peaks.max(), 3.0);
-        
+
         // Add a value that will force erasure of the minimum
-        peaks.push(2.5);  // This should erase 2.0, leaving [1, 3, 2.5]
-        
-        assert_relative_eq!(peaks.min(), 1.0);  // Still 1.0
-        assert_relative_eq!(peaks.max(), 3.0);  // Still 3.0
-        
+        peaks.push(2.5); // This should erase 2.0, leaving [1, 3, 2.5]
+
+        assert_relative_eq!(peaks.min(), 1.0); // Still 1.0
+        assert_relative_eq!(peaks.max(), 3.0); // Still 3.0
+
         // Add another value that will erase the minimum
-        peaks.push(2.7);  // This should erase 1.0, leaving [3, 2.5, 2.7]
-        
+        peaks.push(2.7); // This should erase 1.0, leaving [3, 2.5, 2.7]
+
         // Now the minimum should be recalculated
         assert_relative_eq!(peaks.min(), 2.5);
         assert_relative_eq!(peaks.max(), 3.0);
@@ -271,19 +271,19 @@ mod tests {
     #[test]
     fn test_peaks_stats_after_max_erasure() {
         let mut peaks = Peaks::new(3).unwrap();
-        
+
         // Add values where the maximum will be erased
         peaks.push(1.0);
-        peaks.push(3.0);  // This is the maximum
+        peaks.push(3.0); // This is the maximum
         peaks.push(2.0);
-        
+
         assert_relative_eq!(peaks.min(), 1.0);
         assert_relative_eq!(peaks.max(), 3.0);
-        
+
         // Add values that will eventually force erasure of the maximum
-        peaks.push(1.5);  // This should erase 1.0, leaving [3, 2, 1.5]
-        peaks.push(1.7);  // This should erase 3.0, leaving [2, 1.5, 1.7]
-        
+        peaks.push(1.5); // This should erase 1.0, leaving [3, 2, 1.5]
+        peaks.push(1.7); // This should erase 3.0, leaving [2, 1.5, 1.7]
+
         // Now the maximum should be recalculated
         assert_relative_eq!(peaks.min(), 1.5);
         assert_relative_eq!(peaks.max(), 2.0);
